backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.2
pytest==8.4.2
